Improve BLEMIDI docstrings
